+++ database/migrations/2022_08_01_000015_create_metrika_queries_table.php
@@ -1,0 +1,37 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class CreateMetrikaQueriesTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up(): void
+    {
+        Schema::connection(config('metrika.connection'))->create(config('metrika.tables.queries'), function (Blueprint $table) {
+            // Columns
+            $table->increments('id');
+            $table->longText('query');
+
+            // Indexes
+            $table->unique('query');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down(): void
+    {
+        Schema::connection(config('metrika.connection'))->dropIfExists(config('metrika.tables.queries'));
+    }
+}